requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
cffi==1.17.1
cryptography==44.0.0
Django==5.1.7
mysqlclient==2.2.7
pycparser==2.22
sqlparse==0.5.3
